requirements.txt
aiohttp==3.7.4.post0
aiofiles==0.7.0
Faker==8.13.2
ipwhois==1.2.0
Pillow==8.3.2
requests==2.26.0
scikit-learn==0.24.2